pyproject.toml
[build-system]
requires = [
  "scikit-build-core>=0.10.0",
  "nanobind>=2.2.0",
]
build-backend = "scikit_build_core.build"

[project]
name = "quantithaca"
version = "0.1.0"
description = "QuantIthaca hybrid Python+C++ project"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "QuantIthaca Team" }
]
dependencies = [
    "alembic>=1.18.4",
    "authlib>=1.3.2",
    "django>=5.2.12",
    "django-dnsmanager>=0.2.1",
    "flask>=3.1.3",
    "googlefinance>=0.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pip>=26.0.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.1.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "setuptools>=82.0.1",
    "sqlalchemy>=2.0.0",
    "streamlit>=1.42.0",
    "wheel>=0.46.3",
    "xlwings>=0.34.0",
    "yfinance>=1.2.0",
]

[project.optional-dependencies]
test = []

[tool.scikit-build]
minimum-version = "build-system.requires"
cmake.version = ">=3.25"
cmake.build-type = "Release"
wheel.packages = ["python/quantithaca"]
build-dir = "build/{wheel_tag}"

[tool.pytest.ini_options]
testpaths = ["tests/python"]
markers = [
    "integration: requires TEST_DATABASE_URL (see docker/docker-compose.postgres-test.yml)",
]
